requirements_wsl.txt
annotated-types==0.7.0
click==8.3.0
kafka-python==2.2.15
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
